ref alignment algo in different file
