+++ tests/MavenTests/MavenTests.pro
@@ -19,7 +19,7 @@ QMAKE_CXXFLAGS += -DOMP_PARALLEL
 
 INCLUDEPATH +=  $$top_srcdir/src/core/libmaven  $$top_srcdir/3rdparty/pugixml/src $$top_srcdir/3rdparty/libneural $$top_srcdir/3rdparty/libpls \
 				$$top_srcdir/3rdparty/libcsvparser $$top_srcdir/src/cli/peakdetector $$top_srcdir/3rdparty/libdate $$top_srcdir/3rdparty/libcdfread \
-                $$top_srcdir/3rdparty/obiwarp $$top_srcdir/src/pollyCLI
+                $$top_srcdir/3rdparty/obiwarp $$top_srcdir/src/pollyCLI $$top_srcdir/src/core/libmaven/alignment
 
 
 QMAKE_LFLAGS += -L$$top_builddir/libs/
